Fix active sectors being set to upper storey when on ground floor
 - INVESTIGATION: The issue of walls appearing at strange heights (e.g. in
Lumbridge Castle) is actually not a bug with the way walls are loaded, but is
the result of rendering the upper storeys when indoors (to be fixed later).
The same issue can be seen in the original game when forcing upper storeys to
be rendered.

 - INVESTIGATION: Doors, objects, NPCs, etc. are not normally loaded by the
client, but are sent by the server. Maybe we will do this differently.

 - Additional refactoring (rename variables, add comments)

 - Update docs