Fix inconsistent mouse picking
The mouse picked objects are produced as a direct result of rendering.
Therefore, it is essential that rendering takes place in the correct sequence
within the game loop. To ensure this, we render directly in our main thread
instead of scheduling a repaint in the Swing UI thread.

The game loop contains an inner loop which ticks the game multiple times per
render operation. This change also moves pollInput() outside of this inner
loop, to ensure that the input is only polled once per render.

Later we may consider simplifying the game loop, with one tick per render.